chore: Refactor resonator information retrieval and improve user authentication process
- Remove unused constructor repositories
- Update resonator CSS class
- Refactor resonator template to display retrieved information
- Optimize user authentication process for better performance
