+++ templates/wuwagen.html
@@ -76,7 +76,7 @@
             </div>
             <div class="max-w-6xl mx-auto mt-4 p-2 md:p-0">
                 <p class="text-base md:text-2xl text-white dark:text-gray-200 z-20 relative font-light leading-relaxed">
-                    Retreive informations about the resonator you want from your account, and share the generated build image. (cazzi)
+                    Retreive informations about the resonator you want from your account, and share the generated build image.
                 </p>
             </div>
         </div>
